refactor: move child directories to unique path
