+++ denops/@ddc-source-dictionary/lib/dictionary/external.ts
