+++ denops/@ddc-source-dictionary/lib/dictionary/trie.ts
