+++ denops/@ddc-source-dictionary/lib/kv.ts
