+++ denops/@ddc-source-dictionary/lib/trie.ts
