+++ denops/@ddc-source-dictionary/lib/util.ts
